requirements.txt
Flask==2.0.2
Flask-APScheduler==1.12.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
PyYAML==6.0
snowflake-connector-python==2.7.1